gpu: clarify/fix bit layout of oddly packed formats
rgb10a2, rgb565, and similar packed formats, all follow the convention
that the most significant bit represents the first channel (i.e. red).
This is inconsistent with the usual ordering for byte-aligned planes,
where the first channel (e.g. red) is in the *least* significant byte.

Since our usual convention is to follow the little endian byte-aligned
ordering, the upshot of this is that rgb565 etc. are simply
inconsistent with the "expected" channel order.

Make sure to properly document and implement this inconsistency. I hope
nobody actually uses these formats - but if they do, I'm sure we'll hear
about it.