utils/upload: add new helper, pl_plane_data_align
This makes the job significantly easier of taking something weird like
p010 or yuv420p10 and finding a good "equivalent" GPU-compatible format
it can live in.

This function is intended to be combined with the other helpers, in
particular pl_plane_data_from_mask.

Note that due to limitations of the pl_color_repr struct design, we can
only generate a single `pl_bit_encoding` for the entire sample. This
means that we have to, somewhat arbitrarily, enforce that the
shift/depth is consistent for each component.