colorspace: improve handling of partially set pl_bit_encoding
All of the places in the code that consult `color_depth` or
`sample_depth` now default their values to the other of the two in their
absence.

This makes the result less surprising when e.g. just setting
`color_depth` to 16, and leaving `sample_depth` unset. Previously, this
would effectively cause it to default to 8, rather than 16.