renderer: don't double-apply sample order
The sample order isn't really relevant at all for the pl_renderer,
because it's only meaningful for determining the relationship between
the sample index and its *host representation*.

Consulting this here led to it accidentally being double-swapped for
e.g. BGRA textures, which already swap as part of the component mapping
if needed.