utils/libav: fix signature of pl_upload_avframe
An array of pointers to pointers? Complete rubbish and overkill. An
array of pointers is fine. So what if users have to group their plane
textures together, big deal. It's way more convenient this way, than
having to explicitly separate the plane texture by plane and somehow
provide them separately as an array.

Not even gonna bump the API version for this because 1) nobody uses this
API yet, and 2) it's single-header garbage anyway, so no risk of ABI
break.