shaders/av1: relax luma==0 restriction
After having managed to contrive in my head a scenario in which the luma
plane somehow ends up with luma in an index other than 0 (uploading UYVY
to an rgba8 texture and aliasing it as rg8, the luma information is now
in the G channel), I realized it's a good idea to support this.

It's not really a horrendeous break of API, because the backwards
compatible behaviour is, handily, what you get if you just ignore this
new field's existence.