vulkan: allow disabling page-rounding imported pointers
This is technically a hack, or at the very least, some sort of
ID/UD-type behaviour. It works in practice (for good reasons), but
juuust in case, it would really be a smart idea to allow disabling it.

Also swap the order of the error messages to make sure checks happen in
a more sensible order.