renderer: fix subsampled plane alignment (again)
I never seem to get a break with this logic. Anyway, what happened
*this* time was that I noticed `rx`, `ry` were being incorrectly
calculated (they're supposed to be floats, but I was dividing two ints).

I also realized while working through a paper example that using `rx`
for the plane shifts was incorrect to begin with. Fortunately, for most
real-world content, `rx` and `rrx` are identical. For cases where they
aren't, the earlier answer ended up effectively cancelling out this
error, at the cost of an off-by-one (resulting in tiny, probably
unnoticed chroma misalignment).

The correct logic is to use rrx/rry for both. Fixed, and tested. (In
particular, by making sure a YCbCr image survives being converted to
RGB and back to YCbCr by the renderer)