shaders/colorspace: fix dithering to high depth
The old code tried printf'ing a small float using %f, which resulted in
a rounded literal that was quite far from 1/scale. This led to the image
getting brighter than it should be, globally. The effect was very
significant for higher bit depths.

Fixed this, and improved the precision, by printing exact integer
literals, which get turned into exact float literals by the shader
compiler.