utils/libav: sanitize some illegal metadata combinations
As clarified by @cyanreg, FFmpeg filters are supposed to ignore
obviously "wrong" combinations, in particular RGB + limited range. We
also make an effort to clean up broken sig_peak tagging, since this
exists in practice.

I decided not to sanitize the other paramaters until I encounter a
broken file that demand it, because it's hard to draw the line
otherwise, although we could certainly come up with a few more ideas for
things that are potential candidates. (ICtCp with non-HDR curve? XYZ?)

Also clean up some whitespace because I can't be bothered making a
separate commit for it.