vulkan: decouple samplers from pl_tex_vk
VkSamplers aren't actually tied to any specific image or imageview, so
we can statically create them on pl_gpu creation and just assign the
correct sampler when updating the descriptor.

This reduces the amount of unnecessary objects per texture
creation/destruction by a grand total of one (1).

Sort of pointless at the moment, but the idea moving forwards is to also
decouple the idea of samplers from the `pl_tex` abstraction entirely.