renderer: fix aliasing in corner case
This logic was old and wrong, because params->upscaler and
params->downscaler might point to the same thing, in which case this can
end up being `true` even when downscaling.

Essentially, this would trigger if both upscaler and downscaler are set
to something like bicubic, in which case downscaling would alias.