gpu: add asynchronous texture transfers
This allows us to not only fix some of the shortcomings of pointer-based
transfers, but also make the API easier for users that need to e.g.
deref external images once an operation completes. Due to OpenGL
restrictions, this is dependent on a new GPU capability flag.

Partially implements https://code.videolan.org/videolan/libplacebo/-/issues/123