gpu: nuke pl_buf_pool
I think the status quo is that creating and destroying pl_buf should be
as cheap as possible, so there's really no point in having a separate
abstraction for somehow pooling buffers together. This is a left-over
from when `pl_buf_destroy` on an in-use buffer was undefined behaviour,
which hasn't been the case for some time now.

Removed not only because they're unnecessary, but also because the
unnecessary boilerplate is getting in the way of a planned refactor.