bstr: dramatically increase float precision
It occurred to me that it really makes no sense to limit the precision
here, and it *has* bitten us in a number of ways, and also caused
general headaches in the form of needing to pass some floats as 1.0/%f
to escape the precision loss.

I seem to have completely forgotten that we actually control the float
printing code and can set the preciison to whatever we want. So just set
it as high as reasonable, because this is intended for machine consumption
anyway.

20 digits gives us enough to cover the ~16 digits of the IEEE mantissa
alongside some extra for the order of magnitude (~11 bits). This really
ought to be enough for everyone.