filters: rewrite compute_row
This function was still LGPLv2. Rewrite it to make it our own code.
Incidentally, the new version fixed what was technically a bug in the
old version, namely that the `x` coordinate had the wrong sign. (This,
however, did not matter in practice because `pl_filter_sample` took the
absolute value of the input coordinate anyway)