opengl: implement host pointer imports
This relies on GL_AMD_pinned_memory, which is available on most AMD
devices and most Mesa-based drivers. The usual caveats apply.

Moved a lot of the logic out of vulkan-specific code and into the pl_gpu
wrapper layer. Unfortunately, this required axing the debug option to
disable overmapping. If it causes problems in practice, we probably need
to disable this entire logic anyways.

A minor note of annoyince is that I can't find an easy way of combing
buffer offsets with glDrawArrays. We probably need to update all of the
VAO pointers in this case, which is obnoxious and hard to test. For now,
I've elected to just disable that combination.

Even more annoyingly, GL gives us no way of getting the page size from
the driver. We have to query it with platform-dependent mechanisms...