opengl: fix possible recursive callback loops
If calling a callback calls more OpenGL commands, then gl_poll_callbacks
can be called in an infinite loop, because it will call this same
callback.

Fix this by removing the callback from the list *before* we execute the
callback. This can still recurse, but now it does so safely.