renderer: merge similar planes before heavy operations
This makes most types of planar operations faster, especially
scaling, debanding, AV1 grain application and other plane shaders.
Care needs to be taken to avoid doing this unnecessarily though, to
avoid an FBO roundtrip when there's no point.

Note that we try re-using the same plane texture format if possible, and
only pick as many components for the FBO as we deem necessary. This cuts
down on memory bandwidth.

Closes https://code.videolan.org/videolan/libplacebo/-/issues/26