utils/dav1d: refactor pl_upload_dav1dpicture
I noticed that `film_grain` should not get auto-applied unless
`apply_grain` is disabled in Dav1dSettings by the user. So we need to
make this configurable.

Rather than bikeshed with more variants or parameters, I decided to do
the usual thing and merge it all into a params struct. This solves the
annoying `_and_unref` issue as well, and allows us to be slightly saner
in the case of ambiguously allocated Dav1dPictures.