shaders/colorspace: refactor icc/3dlut API
Renamed from 3DLUT to ICC across the board, and moved to its own
conditionally installed header Also, merged with the internal lcms code,
since it was a bit redundant now.

Mostly to make room for more LUT formats I will be adding, and since
colorspace.h was a bit cluttered.