renderer: avoid unnecessary anamorphic scaling before plane merging
This can happen in the case of slight aspect ratio mismatches between
the target and image, in which case we end up performing fractional
cropping. However, currently, this causes a pathological mismatch
between the refplane and the desired crop, which ends up being
(erroneously) scaled away before plane merging. Not a huge deal, but
waste of performance and worse for quality. Fix it by moving all
subpixel crops to the main scaler.

Note that this means we will get an incorrect scaling ratio in the case
where the plane scaler ends up becoming the main scaler. I don't think
there's a fundamental solution here other than replicating the
`need_fbo` logic from pass_scale_main.

Maybe in general we should move all of the logic for determining which
operations to perform to a common function that runs once, at the
beginning, and caches all results.