demos/plplay: rewrite on top of proper decode/present loop
This threads the decoder loop to make it run asynchronously from the
present loop, so we can fetch frames from it as fast as possible /
required.

We also switch to the new `pl_queue` API, using it to translate frame
timestamps into `pl_frame_mix` structs suitable for interpolation.

This currently defaults on, but I still need to hook it up to the UI.