colorspace: add color temperature adjustment
In theory, this is not quite the right place to do this adjustment,
because it's applied in gamma light. But the alternative fix here, I
think, would be making `pl_renderer` smart enough to apply these
controls on its own. Plus, I don't think this feature is worth the extra
API complexity of somehow moving it to a separate location. The more we
can roll into this struct, the better.

Supporting this, a helper is provided to compute the CIE xy coordinates
for a given color temperature, as well as a function to compute a
chromatic adaptation matrix.