filters: refactor and expand filter presets
Rename this to pl_filter_preset to make the intent clearer, and add an
extra field `description` to provide a friendlier name.

In addition to this change, re-order the list to move the popular
filters up to the top of the list, as well as adding a "none" entry
corresponding to the built-in sampling.

Finally, add pl_frame_mixers to provide frame mixer presets.

Update plplay to use these new fields/structs.