renderer: only force-enable 3DLUT for actual conversions
This logic is not perfect, but good enough for what's essentially a
debug feature. This (conservatively) avoids forcing an ICC for what we
essentially know is a no-op transformation.

Importantly, this somewhat relaxed check is never a false negative.
Also rewrite the logic a bit to reduce symbol clutter.

Fixes https://code.videolan.org/videolan/libplacebo/-/issues/134