ci: delete unused cross files
These are pulled from the CI image anyway. Their presence in the repo
itself makes no sense since all of the information inside it is
environment-specific, meaning it only works on the developer's
machine(tm).

Unfortunately the reality of cross compiling in 2021 is that people will
just have to do it themselves. Which is probably irrelevant anyway
because who the fuck is manually cross-compiling stuff instead of
relying on existing distro tools to do it for them?