libplacebo: introduce the concept of thread safety
Right now, everything is thread unsafe. But that is expected to change.

Also, introduce new helper to assist in creating pthread mutexes, as
well as a refcounting helper. Use the latter in place of the weird mutex
refcount in `pl_ref`.

Finally, delete a left-over refcount field that wasn't actually used for
anything.