utils/queue: properly multi-thread
This is a major refactor with the goal in mind of allowing
`pl_queue_push` to run concurrently with `pl_queue_update`, to bridge
the gap in functionality between the push and pull models of operation.
To avoid overfilling the queue, we add a new blocking variant of
`pl_queue_push`

Notably, this also adds a `uint64_t timeout` field for the `get_frame`
callback to use, something which I had strongly considered doing even
before `pl_queue` was able to interpret it. But now that our blocking
functions interpret it directly even in the non-callback case, there's
no more reason not to add it.

This also adds a `void *user_data` to `struct pl_frame`, since
effectively `pl_frame` is the only way to communicate extra opaque
metadata between threads when using the `pl_queue` API.