demos/plplay: refactor decoder loop
Now that `pl_queue` is thread safe, we can avoid having to queue frames
manually and instead directly push them into the `pl_queue` from the
decoder thread.

In addition to this, rewrite the timing loop in the rendering thread to
set a less aggressive timeout, and avoid advancing the PTS when this
timeout is reached.

This makes `plplay` both simpler and more robust.

As a side note, several PTS-related bugs were fixed, as well as an
aspect ratio bug.