gpu: make thread safe
For the time being, I decided to keep the GPU objects themselves thread
unsafe. I think this is a reasonable tradeoff, and also one that roughly
matches the vulkan threading model. Importantly, the `pl_*_poll` family
of functions are still thread safe, which is good enough for all use
cases I can think of.

This solves exactly that real-world use case of allowing decoder threads
to create their own mapped buffers, which is a large part of the
motivation of making libplacebo thread safe to begin with.

We introduce a new capability to signal thread safety, because OpenGL is
not thread safe (and cannot be, by design of OpenGL).

Closes https://code.videolan.org/videolan/libplacebo/-/issues/68