vulkan/malloc: make thread safe
We accomplish this using one lock to guard the list of heaps (as well as
the list of slabs within that heap), and a separate lock to guard each
slab.

The reason for this split is to avoid having to lock the entire
`vk_malloc` structure inside `vk_malloc_free`, allowing memory from
separate slabs to be freed concurrently and without going through a
single bottleneck.

As an aside, weaken the annoying slab leak check to make it less
annoying, since proper state cleanup in multi-threaded programs is
hard(tm) and leaking memory on program exit is unlikely to matter
anyway.

This check still mainly exists for me, not users. I could justifiably
hide it behind an env var. But whatever, the current is probably good
enough. Users *probably* want to be properly cleaning up after their GPU
resources anyway. Probably.