renderer: completely refactor overlays
The current design was simply lacking for real world purposes, because
apart from VLC, basically every provider of overlays needs to render
multipe sub-parts from the same texture.

I decided to go for a middle ground between flexibility and power.
Technically, I could expose this up to the point of allowing arbitrary
vertex/index buffers, but I simply think that's overkill for most users.

The only scenario in which it's really relevant would be for the nuklear
UI, but users can already custom-roll that themselves if they really
want to. (See `plplay.c`)

Miraculously, backwards compatibility was not actually hard to achieve.
Also, fix a bug where overlays were flipped upside down in sdlimage. I'm
not sure if there's a good fix to this that doesn't rely on the user
remembering `frame.flipped` exists.