utils/frame_queue: relax backwards PTS jump stringency
mpv in particular likes generating small backwards jumps sometimes. In
practice, we can deal with them more or less fine - the only issue is
when we step so far back that we no longer have those frames in the
queue at all. Error in this case. In all other cases, just deal with it.