dispatch: write the correct number of passes in pl_dispatch_save
It's valid for pl_pass_create not to write a cached program. In this
case, pl_dispatch_save should only write the number of passes that have
a program to save, in order to prevent a buffer overread in
pl_dispatch_load.

This also updates pl_shader_tests to handle passes that don't write
a cached program.