log: rename from pl_context
Over the past few years, the use case of `pl_context` has increasingly
shrinked, finally dwindling to "nothing other than logging" with the
introduction of thread safety (and thus the loss of the "thread safety
guard") As such, this being called `pl_context` is really unnecessarily
confusing. It's just a glorified logging subsystem.

Rename it, and make it optional in the process. Also introduce a typedef
to make the new name shorter, as part of a new convention that I will be
switching to moving forwards.

This technically breaks the auto-cleanup behaviour of `pl_shader` and
`pl_filter` but I really don't think those were all that useful or
justified, and I don't think anybody relied on it.