vulkan: properly support -Dvulkan-link=false
When compiling with vulkan linking disabled, we define `vulkan_headers`
as a dependency object that only includes vulkan.h, rather than linking
against libvulkan, and use this for the main libplacebo library. This
allows libplacebo itself to avoid linking against libplacebo altogether
when this option is enabled, as was clearly the intent. The test
framework and demos still use the regular, linked libvulkan.

Supporting this, make sure we compile with VK_NO_PROTOTYPES in this
case, and also fix a bug where we accidentally still directly called the
linked vkCmdCopyImage, and another bug where we forgot to forward
`get_proc_addr` to `pl_vulkan_choose_device`.

Finally, scan for the presence of `vulkan.h` itself and use that as a
fallback for `vulkan_headers`.