alloc: merge pl_*alloc_[ptr_]priv into pl_*alloc_obj
Replaces the four existing variants of alloc/zalloc priv and ptr_priv
(of which only three were defined/used, for some reason) into a single
pair of functions pl_alloc_obj and pl_zalloc_obj that only take a
pointer to the object.

This reduces duplication of the object type names and makes these
functions less verbose in general.