shaders/sampling: use textureGather() at finer granularity
The current code always iterated in 2x2 blocks and only used
textureGather() on blocks aligned to this 2x2 grid. This is needlessly
wasteful, because (for even-sized radii in particular), we can squeeze
in an extra textureGather() block by offsetting them all by one. For
larger radii (radius 4 and up), we can also squeeze in extra blocks on
the "sides".

While this algorithm is still not entirely perfect (ideally we'd use
some square-packing algorithms to maximimize the number of gathered
texels), it's a modest improvement to performance in particular for the
radius 2 case. Radius 3 is mostly unaffected, unfortunately.