utils/frame_queue: properly handle PL_QUEUE_MORE
Don't hard-error on PL_QUEUE_MORE as part of interpolate(), since we can
still return a valid frame mix. The semantics of PL_QUEUE_MORE are that
we should still return as many frames as possible, though we should
inform the user about the fact that it may be truncated.

This should now be working consistently throughout.