utils/frame_queue: switch from ZOH to nearest neighbour
ZOH is actually a somewhat terrible choice of fallback mechanism,
because we encourage users to submit the same presentation PTS as the
source frame PTS. This, combined with floating point uncertainty,
basically guarantees "frame flickering" under ZOH semantics.

Nearest neighbour semantics eliminate this problem, and as such are a
much better fit for the fallback, since it more closely resembles the
result of actually interpolating this frame mix.

Might decide to revisit this, yet again, in the future.