renderer: use nearest-neighbour fallback semantics
Similar to the corresponding commit for the frame_queue, nearest
neighbour semantics are way less fiddly than true ZOH. Basically, think
of it as trying to solve a problem of probability - which frame is most
likely to best overlap with the current frame?

In theory, an even better solution here would be to construct the
`oversample`-like semantics and then pick the frame with the highest
weight. But that's more code to write, and nearest neighbour more or
less gets the right result in practice. (I might revisit this in the
future)