renderer: change semantics for pl_image_mix.num_frames == 1
The current semantics are a bit potentially useless in edge cases
involving only a single frame with PTS far outside the vsync range,
which unfortunately vo_placebo likes to trigger.

In a perfect world I would probably prefer to redesign this entire
abstraction with hindsight, but for the time being it's easier to just
make this small amendment.

Bump the API version mostly as a way to allow vo_placebo to depend on
this new logic (plus the previous fixes).