vulkan: make some pNext chains properly conditional
VkDevice-level pNext chains must not include any pointers to non-enabled
structures.

In theory we also shouldn't do this for VkPhysicalDevice-level
functionality, but because implementations are also required to ignore
unrecognized structs, this means a given VkPhysiclaDevice either
supports the extension (in which case it's allowed to link it) or it
doesn't (in which case it's defined to be ignored).