vulkan: make vulkan 1.1 the minimum required version
For both the VkInstance and the VkDevice. This removes the awkward
dependency on VK_KHR_get_physical_device_properties2, and instead makes
it so we can just directly use vkGetPhysicalDeviceProperties2.

For the user-provided VkInstance, merely assert() this as the case,
since *every* loader on every platform should be capable of providing an
instance at version 1.1. As for the VkDevice itself, insert an extra
error check just to be on the safe side. (And also to prevent users from
accidentally doing something stupid like limiting the api version to
1.0)