utils/frame_queue: fix possible thread race assertion
This is especially easy to trigger if the playback loop gets very stuck
(e.g. moving to a different workspace, on wayland).

Basically the issue is that `get_frame` releases `&p->lock_weak`, which
allows other threads to start pushing frames. So while `get_frame` does
ostensibly do what it claims to do on the surface (gets a frame), it
does not necessarily get exactly *one* frame, but could get 2 or 3 (or
any number).

Despite this, the logic in `advance()` was written as though `get_frame`
still got exactly one frame. Simple solution is to just change the `if`
to a `while`. Shouldn't be terribly inefficient because in practice it's
very unlikely that this will get more than 1 or 2 frames extra.