shaders/sampling: fix textureGather coordinates
Sampling from the base is wrong, because textureGather returns the four
texels surrounding the sampled position. What we actually want to do is
sample the center of the four texels we're interested in.

Note: While we could technically use `pos` here instead of calculating
the position dynamically, doing so seems to result in 1 pixel black
seams when combining it with some sampler modes, probably because of the
near-undefined behavior on the exact edge between pixels.

Maybe we could enable this conditionally only if we know the sampler is
set to nearest mode. But for now, just play it safe.

No change in performance. (Thankfully)