colorspace: add more pure power gamma functions
Truthfully the only motivation for this commit is that I want to have
PL_COLOR_TRC_GAMMA24 exist for backwards compatibility with the legacy
handling of (non-black-scaled) BT.1886.

In theory we could also probably do something smarter like encoding the gamma
value into the enum value itself, but that can also be done later and I
don't want this commit to become that invasive. (And it sets a precedent
I don't like with regards to how to handle unknown enum members in
switch statements)