shaders/sampling: use specialization constants
Use constants instead of literals across the board.

For the EWA shader in particular, the ratio-dependent `iw` and `ih` have
also been turned into constants, as well as the filter radius, which now
allows the EWA shader to be re-specialized to any required output size
without triggering a recompilation. (Except when requiring more samples
for anti-aliasing)