shaders: allow lifting constants to variables
This can remove the recompilation overhead from programs such as
`plplay` which expose lots of settings as freely adjustable GUI sliders.
Normally, this would trigger a recompilation on every frame. Lifting
those constants to variables improves this.

Care needs to be taken not to lift constants that will be used as the
inputs of arrays, such as with the polar sampling code. Also, we now
need to avoid the `const` keyword on things that ingest `sh_const`.

Add an API for this to pl_dispatch and pl_renderer, respectively.