renderer: mix in the rendered frame's native color space
This is far better than ping-ponging between this and `image.color`,
especially because it avoids e.g. unnecessarily adapting the black point
back and forth between it and linear light. In the ideal case, this
makes it a no-op.

It also means we now effectively perform frame blending in linear light,
which I think is a net gain anyways, especially because it also avoids
needing to unlinearize every single frame in the mixing shader (though
one could in theory also easily fix that problem by moving it to the
previous step).

Also fixes an issue where this adaptation resulted in incorrect results,
although truthfully I'm not entirely sure why.