demos: color manage the nuklear UI
Otherwise, on HDR / wide gamut swapchains, this results in hilariously
broken output.

As a side note, this may be somewhat undesirable for the background
color selector, since the actual chosen color in the UI will not
correspond to the color as seen on-screen. Maybe we should have
pl_renderer treat the background color as sRGB too. That would make it
impossible to have super-bright HDR backgrounds though. Does anybody
care about that?