vulkan/malloc: rename 'heap' to 'pool' to avoid clashing terms
The mismatch between vk_heap and actual vulkan heaps is pretty
confusing even to me. Rename our 'heap' concept to 'pool' in order to
just give it a different name.

Also, keep track of each pool's index (which is not expected to change)
and add an extra error message to failed allocations indicating the
exact number of bytes that we failed serving, to better diagnose issues
with the space map.