vulkan/command: fix thread safety issues
Releasing the mutex while blocking can invalidate the state after the
blocking command completes. We need to re-check whether this command is
still relevant, or if another thread got to it first.

Note that, with the way this is written, this can result in
double-waiting on the same fence, which is harmless in Vulkan.