opengl: properly disable interfering state
Due to OpenGL being annoying and stateful, it's possible for
applications using libplacebo to leak state (in particular, face culling
and depth testing) that interferes with libplacebo's own rendering.

Due to reasons of annoying annoyingness, it's slightly non-trivial to
make libplacebo generate vertices in clockwise order (reason: we fix the
order of texture coordinates in vertex order ahead of time, so if the
position is suddenly flipped w.r.t the corresponding texture coordinate,
we'd need non-trivial communication across vertices to undo that).

Easiest solution is to just disable face culling for libplacebo render
passes, which makes it match the behavior of vulkan. This is fine since
we don't ever care about face culling - indeed, it's just an unnecessary
slowdown.